esim mux_4to2_encoder
This project focuses on designing a 4-to-2 Encoder using a Multiplexer-based approach with Skywater 130nm CMOS technology. The primary objective was to optimize the encoder's design for efficient digital logic implementation, focusing on reduced component usage and improved power performance. The design was modeled, synthesized, and pre-layout simulated using eSim and NgSpice.